+++ webflux/src/main/java/be/webflux/services/IOrganisationService.java
@@ -1,14 +1,16 @@
 package be.webflux.services;
 
+import java.math.BigInteger;
+
 import be.webflux.domain.Organisation;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface IOrganisationService {
 	 void create(Organisation o);
-	 Mono<Organisation> findById(Long id);
+	 Mono<Organisation> findById(BigInteger id);
 	 Flux<Organisation> findByName(String name);
 	 Flux<Organisation> findAll();
 	 Mono<Organisation> update(Organisation o);
-	 Mono<Void> delete(Long id);
+	 Mono<Void> delete(BigInteger id);
 }
